Implemented most of the changes to the database proposed by Lars Huttar.
I have excluded the changes aimed at reducing the amount of information
given (mainly for quest artifacts). Before implementing these changes
we would need to provide another mechanism that the more detail information
could be discovered with. One possibility discussed in slashem-devel is
for the oracle to provide this information for quest artifacts in the
hero's inventory.

Lar's original patch can be found in the slashem-devel archives
(mailed on Sun, 30 Sep 2001).