Fix a problem with writing profile values out that aren't recognized and which contain an empty vector.
It's possible to hit this error by using a CVS version of gtkhack to
create a profile containing a connections profile with no flags set
and then attempting to use the same profile with a proxified version
of the GTK interface (which doesn't interpret "connections"). It's
not possible to hit this in a released version of Slash'EM so I'm
not bothering to log a proper bug report.